chore(base): Rename `rooms` to `room`.
This patch renames the `rooms` module into `room`. It contains a single
kind of `Room`.
