+++ crates/matrix-sdk-base/src/lib.rs
@@ -34,7 +34,7 @@ pub mod latest_event;
 pub mod media;
 pub mod notification_settings;
 mod response_processors;
-mod rooms;
+mod room;
 
 pub mod read_receipts;
 pub mod sliding_sync;
@@ -54,7 +54,7 @@ pub use http;
 #[cfg(feature = "e2e-encryption")]
 pub use matrix_sdk_crypto as crypto;
 pub use once_cell;
-pub use rooms::{
+pub use room::{
     apply_redaction, EncryptionState, Room, RoomCreateWithCreatorEventContent, RoomDisplayName,
     RoomHero, RoomInfo, RoomInfoNotableUpdate, RoomInfoNotableUpdateReasons, RoomMember,
     RoomMembersUpdate, RoomMemberships, RoomState, RoomStateFilter,
