+++ crates/matrix-sdk-base/src/room/display_name.rs
