+++ crates/matrix-sdk-base/src/room/encryption.rs
