+++ crates/matrix-sdk-base/src/room/latest_event.rs
