+++ crates/matrix-sdk-base/src/room/state.rs
