+++ crates/matrix-sdk-base/src/store/mod.rs
@@ -60,7 +60,7 @@ use tracing::warn;
 use crate::{
     deserialized_responses::DisplayName,
     event_cache::store as event_cache_store,
-    rooms::{RoomInfo, RoomInfoNotableUpdate, RoomState},
+    room::{RoomInfo, RoomInfoNotableUpdate, RoomState},
     MinimalRoomMemberEvent, Room, RoomStateFilter, SessionMeta,
 };
